# docs/swagger.yaml
basePath: /
definitions:
...
  genproto.LoginReq:
    properties:
      Password:
        type: string
      UserName:
        type: string
    type: object
  genproto.Post:
    properties:
      Body:
        type: string
      CategoryId:
        type: string
      PostId:
        type: string
      Title:
        type: string
      UserId:
        type: string
    type: object
...
  genproto.Posts:
    properties:
      Posts:
        items:
          $ref: '#/definitions/genproto.Post'
        type: array
    type: object
  genproto.Tag:
    properties:
      Name:
        type: string
      TagId:
        type: string
    type: object
  genproto.TagList:
    properties:
      Tags:
        items:
          $ref: '#/definitions/genproto.Tag'
        type: array
    type: object
  genproto.Token:
    properties:
      AccessToken:
        type: string
      RefreshToken:
        type: string
    type: object
  genproto.User:
    properties:
      Email:
        type: string
      Password:
        type: string
      UserId:
        type: string
      UserName:
        type: string
    type: object
  genproto.UserModel:
    properties:
      Email:
        type: string
      Password:
        type: string
      UserName:
        type: string
    type: object
  genproto.Users:
    properties:
      Users:
        items:
          $ref: '#/definitions/genproto.User'
        type: array
    type: object
info:
  contact: {}
  title: Online Forum API Documentation
paths:
...
  /admin/user/{id}:
    delete:
      consumes:
      - application/json
      description: This api Deletes User
      parameters:
      - description: user id to be deleted
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: user deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Deletes User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: This api Updates User
      parameters:
      - description: user id to be updated
        in: path
        name: id
        required: true
        type: string
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/genproto.UserModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.User'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Updates User
      tags:
      - User
  /admin/user/all:
    get:
      consumes:
      - application/json
      description: This api Get all Users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Users'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Gets all User
      tags:
      - User
...
  /tag/{id}:
    delete:
      consumes:
      - application/json
      description: This api Deletes Tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: tag deleted successfully!
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Deletes Tag
      tags:
      - Tag
    put:
      consumes:
      - application/json
      description: This api Updates Tag
      parameters:
      - description: Tag Name
        in: query
        name: name
        required: true
        type: string
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Tag'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Updates Tag
      tags:
      - Tag
  /tag/{id}/posts:
    get:
      consumes:
      - application/json
      description: This api Gets Post By Tag
      parameters:
      - description: Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Tag'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets Post By Tag
      tags:
      - Tag
  /tag/all:
    get:
      consumes:
      - application/json
      description: This api Gets All Tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.TagList'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets All Tags
      tags:
      - Tag
  /tag/create:
    post:
      consumes:
      - application/json
      description: This api create Tag
      parameters:
      - description: Tag Name
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.Tag'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Creates Tag
      tags:
      - Tag
  /tag/get/{name}:
    get:
      consumes:
      - application/json
      description: This api Gets Tag by Name
      parameters:
      - description: Tag Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Tag'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets Tag by Name
      tags:
      - Tag
  /tag/popular:
    get:
      consumes:
      - application/json
      description: This api Gets Popular Tags
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.TagList'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Gets Popular Tags
      tags:
      - Tag
  /user/{username}:
    get:
      consumes:
      - application/json
      description: This api Gets User info
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.User'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Gets User info
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: This api Logins User
      parameters:
      - description: login credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/genproto.LoginReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.Token'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Logins User
      tags:
      - User
  /user/posts/{id}:
    get:
      consumes:
      - application/json
      description: This api Gets Posts of User
      parameters:
      - description: User Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/genproto.Posts'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Gets Posts of User
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      description: This api create User
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/genproto.UserModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/genproto.User'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Creates User
      tags:
      - User
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"